requirements.txt
Django==3.2.14
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-redis==5.2.0
ipython==8.4.0
mysqlclient==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
bcrypt==3.2.0
drf-yasg==1.20.0
cffi==1.15.1
cryptography==37.0.4
redis==4.3.4